chore(web): remove tiptap v1 residual dep, add dompurify + react-virtual, unify to npm
