+++ web/package.json
@@ -11,12 +11,14 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.2.2",
+    "@tanstack/react-virtual": "^3.13.24",
     "@tiptap/extension-code-block-lowlight": "^3.20.6",
     "@tiptap/extension-placeholder": "^3.20.6",
     "@tiptap/extension-task-item": "^3.20.6",
     "@tiptap/extension-task-list": "^3.20.6",
     "@tiptap/react": "^3.20.6",
     "@tiptap/starter-kit": "^3.20.6",
+    "dompurify": "^3.4.4",
     "highlight.js": "^11.11.1",
     "lowlight": "^3.3.0",
     "mermaid": "^11.13.0",
@@ -25,11 +27,11 @@
     "react-markdown": "^10.1.0",
     "remark-gfm": "^4.0.1",
     "tailwindcss": "^4.2.2",
-    "tiptap": "^1.32.2",
     "tiptap-markdown": "^0.9.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.4",
+    "@types/dompurify": "^3.0.5",
     "@types/node": "^24.12.0",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
