+++ web/src/components/SettingsPanel.tsx
@@ -1,7 +1,14 @@
 import { useEffect, useState, useCallback } from 'react'
 
 import type { PresetCommand } from '../types'
-import RunnerPanel from './RunnerPanel'
+import type { TabId, ToastState } from './settings/shared'
+import { TABS } from './settings/shared'
+import AppearanceTab from './settings/AppearanceTab'
+import SessionsTab from './settings/SessionsTab'
+import PresetsTab from './settings/PresetsTab'
+import LLMTab from './settings/LLMTab'
+import RunnerTab from './settings/RunnerTab'
+import MarketTab from './settings/MarketTab'
 
 interface SettingsPanelProps {
   open: boolean
@@ -10,586 +17,16 @@ interface SettingsPanelProps {
   onPresetsChange?: (presets: PresetCommand[]) => void
 }
 
-type Theme = 'dark' | 'light'
-type FontSize = 'small' | 'medium' | 'large'
-type Language = 'zh-CN' | 'en'
-
-const FONT_SIZE_MAP: Record<FontSize, string> = {
-  small: '14px',
-  medium: '16px',
-  large: '18px',
-}
-
-interface UserSettings {
-  theme: Theme
-  font_size: FontSize
-  nickname: string
-  language: Language
-  preset_commands?: string
-}
-
-const DEFAULT_SETTINGS: UserSettings = {
-  theme: 'dark',
-  font_size: 'medium',
-  nickname: '',
-  language: 'zh-CN',
-}
-
-// localStorage fallback keys
-const LS_KEYS: Record<string, string> = {
-  theme: 'xbot-theme',
-  font_size: 'xbot-font-size',
-  nickname: 'xbot-nickname',
-  language: 'xbot-language',
-}
-
-function lsGet<K extends keyof UserSettings>(key: K, fallback: UserSettings[K]): UserSettings[K] {
-  const raw = localStorage.getItem(LS_KEYS[key])
-  return (raw as UserSettings[K]) || fallback
-}
-
-function lsSet<K extends keyof UserSettings>(key: K, value: UserSettings[K]) {
-  localStorage.setItem(LS_KEYS[key], value as string)
-}
-
-async function fetchSettings(): Promise<UserSettings & Record<string, string>> {
-  try {
-    const resp = await fetch('/api/settings')
-    const data = await resp.json()
-    if (data.ok && data.settings) {
-      return {
-        theme: (data.settings.theme as Theme) || lsGet('theme', DEFAULT_SETTINGS.theme),
-        font_size: (data.settings.font_size as FontSize) || lsGet('font_size', DEFAULT_SETTINGS.font_size),
-        nickname: data.settings.nickname || lsGet('nickname', DEFAULT_SETTINGS.nickname),
-        language: (data.settings.language as Language) || lsGet('language', DEFAULT_SETTINGS.language),
-        preset_commands: data.settings.preset_commands,
-        // Agent settings
-        context_mode: data.settings.context_mode || 'auto',
-        max_iterations: data.settings.max_iterations || '2000',
-        max_concurrency: data.settings.max_concurrency || '3',
-        max_context_tokens: data.settings.max_context_tokens || '200000',
-        max_output_tokens: data.settings.max_output_tokens || '8192',
-        thinking_mode: data.settings.thinking_mode ?? '',
-        enable_auto_compress: data.settings.enable_auto_compress ?? 'true',
-        enable_stream: data.settings.enable_stream ?? 'true',
-        enable_masking: data.settings.enable_masking ?? 'true',
-      }
-    }
-  } catch {
-    // Server unreachable — use localStorage fallback
-  }
-  return {
-    theme: lsGet('theme', DEFAULT_SETTINGS.theme),
-    font_size: lsGet('font_size', DEFAULT_SETTINGS.font_size),
-    nickname: lsGet('nickname', DEFAULT_SETTINGS.nickname),
-    language: lsGet('language', DEFAULT_SETTINGS.language),
-  }
-}
-
-async function saveSettings(settings: Partial<UserSettings>): Promise<boolean> {
-  try {
-    const resp = await fetch('/api/settings', {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ settings }),
-    })
-    const data = await resp.json()
-    return data.ok === true
-  } catch {
-    return false
-  }
-}
-
-type TabId = 'appearance' | 'sessions' | 'presets' | 'llm' | 'runner' | 'market'
-
-interface MarketEntry {
-  id: number
-  type: string
-  name: string
-  description: string
-  author: string
-  created_at: string
-  installed: boolean
-}
-
-interface MyMarketEntry {
-  name: string
-  type: string
-  description: string
-  published: boolean
-}
-
-interface SessionInfo {
-  id: string
-  type: string       // "main" | "subagent"
-  label: string
-  role?: string
-  instance?: string
-  running?: boolean
-  preview?: string
-  members?: string
-}
-
-interface SessionMessage {
-  role: string
-  content: string
-}
-
-const TABS: { id: TabId; label: string; icon: string }[] = [
-  { id: 'appearance', label: '外观', icon: '🎨' },
-  { id: 'sessions', label: '会话', icon: '💬' },
-  { id: 'presets', label: '快捷指令', icon: '⚡' },
-  { id: 'llm', label: 'LLM', icon: '🧠' },
-  { id: 'runner', label: 'Runner', icon: '🖥️' },
-  { id: 'market', label: '市场', icon: '🏪' },
-]
-
-// ── LLM Config types ──
-
-interface LLMConfig {
-  provider: string
-  base_url: string
-  model: string
-  models: string[]
-  is_global: boolean
-}
-
-const PROVIDER_OPTIONS = [
-  { value: 'openai', label: 'OpenAI (GPT / o-series)' },
-  { value: 'anthropic', label: 'Anthropic (Claude)' },
-]
-
 export default function SettingsPanel({ open, onClose, onNicknameChange, onPresetsChange }: SettingsPanelProps) {
   const [activeTab, setActiveTab] = useState<TabId>('appearance')
-  const [theme, setTheme] = useState<Theme>(() => lsGet('theme', DEFAULT_SETTINGS.theme))
-  const [fontSize, setFontSize] = useState<FontSize>(() => lsGet('font_size', DEFAULT_SETTINGS.font_size))
-  const [nickname, setNickname] = useState<string>(() => lsGet('nickname', DEFAULT_SETTINGS.nickname))
-  const [language, setLanguage] = useState<Language>(() => lsGet('language', DEFAULT_SETTINGS.language))
   const [saving, setSaving] = useState(false)
 
-  const [toast, setToast] = useState<{ id: number; message: string; type: 'info' | 'error' | 'success' } | null>(null)
+  const [toast, setToast] = useState<ToastState | null>(null)
   const showToast = useCallback((message: string, type: 'info' | 'error' | 'success' = 'info') => {
     const id = Date.now()
     setToast({ id, message, type })
     setTimeout(() => setToast(null), 2500)
   }, [])
-
-  const [marketType, setMarketType] = useState<'agent' | 'skill'>('agent')
-  const [marketSubTab, setMarketSubTab] = useState<'browse' | 'mine'>('browse')
-  const [marketEntries, setMarketEntries] = useState<MarketEntry[]>([])
-  const [myMarketEntries, setMyMarketEntries] = useState<MyMarketEntry[]>([])
-  const [marketLoading, setMarketLoading] = useState(false)
-
-  // Preset commands state
-  const [presetList, setPresetList] = useState<PresetCommand[]>([])
-  const [editingPreset, setEditingPreset] = useState<PresetCommand | null>(null)
-  const [presetSaving, setPresetSaving] = useState(false)
-
-  // LLM config state
-  const [llmConfig, setLlmConfig] = useState<LLMConfig | null>(null)
-  const [llmConfigLoading, setLlmConfigLoading] = useState(false)
-  const [llmSaving, setLlmSaving] = useState(false)
-  const [llmMaxContext, setLlmMaxContext] = useState<number>(0)
-  const [llmMaxContextSaving, setLlmMaxContextSaving] = useState(false)
-
-  // Agent settings state (user-scoped, stored in user_settings table)
-  const [_agentSettings, setAgentSettings] = useState<Record<string, string>>({
-    context_mode: 'auto',
-    max_iterations: '2000',
-    max_concurrency: '3',
-    max_context_tokens: '200000',
-    max_output_tokens: '8192',
-    thinking_mode: '',
-    enable_auto_compress: 'true',
-    enable_stream: 'true',
-    enable_masking: 'true',
-    language: '',
-  })
-
-  // Sessions state
-  const [sessions, setSessions] = useState<SessionInfo[]>([])
-  const [sessionsLoading, setSessionsLoading] = useState(false)
-  const [selectedSession, setSelectedSession] = useState<SessionInfo | null>(null)
-  const [sessionMessages, setSessionMessages] = useState<SessionMessage[]>([])
-  const [sessionMessagesLoading, setSessionMessagesLoading] = useState(false)
-
-  const [llmFormProvider, setLlmFormProvider] = useState('openai')
-  const [llmFormBaseUrl, setLlmFormBaseUrl] = useState('')
-  const [llmFormApiKey, setLlmFormApiKey] = useState('')
-  const [llmFormModel, setLlmFormModel] = useState('')
-  const [llmError, setLlmError] = useState('')
-  const [llmFormMaxOutputTokens, setLlmFormMaxOutputTokens] = useState('')
-  const [llmFormThinkingMode, setLlmFormThinkingMode] = useState('auto')
-
-  // Load settings from server on mount
-  useEffect(() => {
-    if (!open) return
-    fetchSettings().then((s) => {
-      setTheme(s.theme as Theme)
-      setFontSize(s.font_size as FontSize)
-      setNickname(s.nickname)
-      setLanguage(s.language as Language)
-      // Load presets from the same response
-      if (s.preset_commands) {
-        try {
-          const parsed = JSON.parse(s.preset_commands)
-          if (Array.isArray(parsed)) setPresetList(parsed)
-        } catch { /* ignore */ }
-      }
-      // Load agent settings
-      setAgentSettings({
-        context_mode: s.context_mode || 'auto',
-        max_iterations: s.max_iterations || '2000',
-        max_concurrency: s.max_concurrency || '3',
-        max_context_tokens: s.max_context_tokens || '200000',
-        max_output_tokens: s.max_output_tokens || '8192',
-        thinking_mode: s.thinking_mode ?? '',
-        enable_auto_compress: s.enable_auto_compress ?? 'true',
-        enable_stream: s.enable_stream ?? 'true',
-        enable_masking: s.enable_masking ?? 'true',
-        language: s.language || '',
-      })
-    })
-  }, [open])
-
-  // Apply theme
-  useEffect(() => {
-    document.documentElement.setAttribute('data-theme', theme)
-    lsSet('theme', theme)
-  }, [theme])
-
-  // Apply font size
-  useEffect(() => {
-    document.documentElement.style.setProperty('--xbot-font-size', FONT_SIZE_MAP[fontSize])
-    lsSet('font_size', fontSize)
-  }, [fontSize])
-
-  // Persist nickname locally
-  useEffect(() => {
-    lsSet('nickname', nickname)
-  }, [nickname])
-
-  // Persist language locally
-  useEffect(() => {
-    lsSet('language', language)
-  }, [language])
-
-  // Fetch LLM config when tab is opened
-  const fetchLLMConfig = useCallback(async () => {
-    setLlmConfigLoading(true)
-    setLlmError('')
-    try {
-      const resp = await fetch('/api/llm-config')
-      const data = await resp.json()
-      if (data.ok) {
-        setLlmConfig({
-          provider: data.provider,
-          base_url: data.base_url,
-          model: data.model,
-          models: data.models || [],
-          is_global: !!data.is_global,
-        })
-        setLlmMaxContext(data.max_context || 0)
-      } else {
-        setLlmConfig(null)
-      }
-    } catch {
-      setLlmError('获取配置失败')
-    }
-    setLlmConfigLoading(false)
-  }, [])
-
-  useEffect(() => {
-    if (open && activeTab === 'llm') fetchLLMConfig()
-  }, [open, activeTab, fetchLLMConfig])
-
-  const handleSave = useCallback(async (updates: Partial<UserSettings>) => {
-    setSaving(true)
-    const ok = await saveSettings(updates)
-    setSaving(false)
-    if (ok) {
-      showToast('设置已保存', 'success')
-    } else {
-      showToast('保存失败，请重试', 'error')
-    }
-  }, [showToast])
-
-  // Preset commands CRUD
-  const savePresets = useCallback(async (list: PresetCommand[]) => {
-    setPresetSaving(true)
-    const sorted = [...list].sort((a, b) => a.sort - b.sort)
-    const ok = await saveSettings({ preset_commands: JSON.stringify(sorted) })
-    if (ok) {
-      setPresetList(sorted)
-      onPresetsChange?.(sorted)
-    }
-    setPresetSaving(false)
-    return ok
-  }, [onPresetsChange])
-
-  const handlePresetAdd = useCallback(() => {
-    setEditingPreset({
-      id: Math.random().toString(36).slice(2, 10) + Date.now().toString(36),
-      label: '',
-      icon: '⚡',
-      content: '',
-      fill: false,
-      sort: presetList.length,
-    })
-  }, [presetList.length])
-
-  const handlePresetSave = useCallback(async (preset: PresetCommand) => {
-    const exists = presetList.find(p => p.id === preset.id)
-    const newList = exists
-      ? presetList.map(p => p.id === preset.id ? preset : p)
-      : [...presetList, preset]
-    const ok = await savePresets(newList)
-    if (ok) setEditingPreset(null)
-  }, [presetList, savePresets])
-
-  const handlePresetDelete = useCallback(async (id: string) => {
-    if (!confirm('确认删除此快捷指令？')) return
-    const newList = presetList
-      .filter(p => p.id !== id)
-      .map((p, i) => ({ ...p, sort: i }))
-    await savePresets(newList)
-  }, [presetList, savePresets])
-
-  const handlePresetMove = useCallback(async (id: string, direction: 'up' | 'down') => {
-    const sorted = [...presetList].sort((a, b) => a.sort - b.sort)
-    const idx = sorted.findIndex(p => p.id === id)
-    if (idx < 0) return
-    const swapIdx = direction === 'up' ? idx - 1 : idx + 1
-    if (swapIdx < 0 || swapIdx >= sorted.length) return
-    const temp = sorted[idx].sort
-    sorted[idx] = { ...sorted[idx], sort: sorted[swapIdx].sort }
-    sorted[swapIdx] = { ...sorted[swapIdx], sort: temp }
-    await savePresets(sorted)
-  }, [presetList, savePresets])
-
-  // LLM config actions
-  const handleLLMAdd = useCallback(async () => {
-    if (!llmFormBaseUrl.trim() || !llmFormApiKey.trim()) {
-      setLlmError('Base URL 和 API Key 为必填项')
-      return
-    }
-    setLlmSaving(true)
-    setLlmError('')
-    try {
-      const resp = await fetch('/api/llm-config', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          provider: llmFormProvider,
-          base_url: llmFormBaseUrl.trim(),
-          api_key: llmFormApiKey.trim(),
-          model: llmFormModel.trim(),
-          max_output_tokens: llmFormMaxOutputTokens ? parseInt(llmFormMaxOutputTokens) : 0,
-          thinking_mode: llmFormThinkingMode,
-        }),
-      })
-      const data = await resp.json()
-      if (data.ok) {
-        setLlmFormBaseUrl('')
-        setLlmFormApiKey('')
-        setLlmFormModel('')
-        setLlmFormMaxOutputTokens('')
-        setLlmFormThinkingMode('auto')
-        await fetchLLMConfig()
-      } else {
-        setLlmError(data.error || '保存失败')
-      }
-    } catch {
-      setLlmError('网络错误')
-    }
-    setLlmSaving(false)
-  }, [llmFormProvider, llmFormBaseUrl, llmFormApiKey, llmFormModel, llmFormMaxOutputTokens, llmFormThinkingMode, fetchLLMConfig])
-
-  const handleLLMSetModel = useCallback(async (model: string) => {
-    setLlmSaving(true)
-    setLlmError('')
-    try {
-      const resp = await fetch('/api/llm-config/model', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ model }),
-      })
-      const data = await resp.json()
-      if (data.ok) {
-        await fetchLLMConfig()
-      } else {
-        setLlmError(data.error || '切换模型失败')
-      }
-    } catch {
-      setLlmError('网络错误')
-    }
-    setLlmSaving(false)
-  }, [fetchLLMConfig])
-
-  const handleLLMDelete = useCallback(async () => {
-    if (!confirm('确认删除个人 LLM 配置？删除后将恢复使用系统默认模型。')) return
-    setLlmSaving(true)
-    setLlmError('')
-    try {
-      const resp = await fetch('/api/llm-config', { method: 'DELETE' })
-      const data = await resp.json()
-      if (data.ok) {
-        await fetchLLMConfig()
-        // Clear form too
-        setLlmFormBaseUrl('')
-        setLlmFormApiKey('')
-        setLlmFormModel('')
-        setLlmFormMaxOutputTokens('')
-        setLlmFormThinkingMode('auto')
-      } else {
-        setLlmError(data.error || '删除失败')
-      }
-    } catch {
-      setLlmError('网络错误')
-    }
-    setLlmSaving(false)
-  }, [])
-
-  const handleLLMMaxContextSave = useCallback(async () => {
-    setLlmMaxContextSaving(true)
-    setLlmError('')
-    try {
-      const resp = await fetch('/api/llm-max-context', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ max_context: llmMaxContext }),
-      })
-      const data = await resp.json()
-      if (!data.ok) {
-        setLlmError(data.error || '保存失败')
-      }
-    } catch {
-      setLlmError('网络错误')
-    }
-    setLlmMaxContextSaving(false)
-  }, [llmMaxContext])
-
-
-  // Agent settings save handler
-
-  // Market functions
-  const loadMarket = useCallback(async () => {
-    setMarketLoading(true)
-    try {
-      const resp = await fetch(`/api/market?type=${marketType}&limit=50`)
-      const data = await resp.json()
-      if (data.ok) setMarketEntries(data.entries || [])
-    } catch { /* ignored */ }
-    setMarketLoading(false)
-  }, [marketType])
-
-  const handleInstall = useCallback(async (entry: MarketEntry) => {
-    try {
-      const resp = await fetch('/api/market/install', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ type: entry.type, id: entry.id }),
-      })
-      const data = await resp.json()
-      if (data.ok) loadMarket()
-    } catch { /* ignored */ }
-  }, [loadMarket])
-
-  const handleUninstall = useCallback(async (entry: MarketEntry) => {
-    try {
-      const resp = await fetch('/api/market/uninstall', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ type: entry.type, name: entry.name }),
-      })
-      const data = await resp.json()
-      if (data.ok) loadMarket()
-    } catch { /* ignored */ }
-  }, [loadMarket])
-
-  const loadMyMarket = useCallback(async () => {
-    setMarketLoading(true)
-    try {
-      const resp = await fetch(`/api/market/my?type=${marketType}`)
-      const data = await resp.json()
-      if (data.ok) setMyMarketEntries(data.entries || [])
-    } catch { /* ignored */ }
-    setMarketLoading(false)
-  }, [marketType])
-
-  const handlePublish = useCallback(async (entry: MyMarketEntry) => {
-    try {
-      const resp = await fetch('/api/market/publish', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ type: entry.type, name: entry.name }),
-      })
-      const data = await resp.json()
-      if (data.ok) {
-        setMyMarketEntries(prev => prev.map(e =>
-          e.name === entry.name && e.type === entry.type ? { ...e, published: true } : e
-        ))
-      }
-    } catch { /* ignored */ }
-  }, [])
-
-  const handleUnpublish = useCallback(async (entry: MyMarketEntry) => {
-    try {
-      const resp = await fetch('/api/market/unpublish', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ type: entry.type, name: entry.name }),
-      })
-      const data = await resp.json()
-      if (data.ok) {
-        setMyMarketEntries(prev => prev.map(e =>
-          e.name === entry.name && e.type === entry.type ? { ...e, published: false } : e
-        ))
-      }
-    } catch { /* ignored */ }
-  }, [])
-
-  // Load market when tab is opened
-  useEffect(() => {
-    if (open && activeTab === 'market') {
-      if (marketSubTab === 'browse') loadMarket()
-      else loadMyMarket()
-    }
-  }, [open, activeTab, marketSubTab, loadMarket, loadMyMarket])
-
-  // Load sessions when tab is opened
-  const fetchSessions = useCallback(async () => {
-    setSessionsLoading(true)
-    setSelectedSession(null)
-    setSessionMessages([])
-    try {
-      const resp = await fetch('/api/sessions')
-      const data = await resp.json()
-      if (data.ok) setSessions(data.rooms || [])
-    } catch { /* ignored */ }
-    setSessionsLoading(false)
-  }, [])
-
-  useEffect(() => {
-    if (open && activeTab === 'sessions') fetchSessions()
-  }, [open, activeTab, fetchSessions])
-
-  // Load session messages when a session is selected
-  const fetchSessionMessages = useCallback(async (session: SessionInfo) => {
-    setSessionMessagesLoading(true)
-    setSessionMessages([])
-    try {
-      const resp = await fetch(`/api/sessions/messages?id=${encodeURIComponent(session.id)}`)
-      const data = await resp.json()
-      if (data.ok) setSessionMessages(data.messages || [])
-    } catch { /* ignored */ }
-    setSessionMessagesLoading(false)
-  }, [])
-
-  // Load session messages when a session is selected
-  useEffect(() => {
-    if (selectedSession) fetchSessionMessages(selectedSession)
-  }, [selectedSession, fetchSessionMessages])
 
   // Close on Escape
   useEffect(() => {
@@ -603,11 +40,6 @@ export default function SettingsPanel({ open, onClose, onNicknameChange, onPrese
 
   if (!open) return null
 
-  const sectionClass = 'settings-section'
-  const sectionTitleClass = 'settings-section-title'
-
-  const providerLabel = PROVIDER_OPTIONS.find(p => p.value === llmConfig?.provider)?.label || llmConfig?.provider
-
   return (
     <>
       {/* Backdrop */}
@@ -616,12 +48,12 @@ export default function SettingsPanel({ open, onClose, onNicknameChange, onPrese
         onClick={onClose}
       />
       {/* Panel */}
-      <div className="settings-panel">
+      <div className="settings-panel" role="dialog" aria-modal="true" aria-label="设置">
         <div className="flex items-center justify-between">
           <h2 className="text-lg font-bold text-white">⚙️ 设置</h2>
           <div className="flex items-center gap-2">
             {saving && <span className="text-xs text-slate-500">保存中...</span>}
-            <button className="settings-close-btn text-sm" onClick={onClose}>✕</button>
+            <button className="settings-close-btn text-sm" onClick={onClose} aria-label="关闭设置">✕</button>
           </div>
         </div>
 
@@ -645,668 +77,28 @@ export default function SettingsPanel({ open, onClose, onNicknameChange, onPrese
         {/* Scrollable content area */}
         <div style={{ flex: 1, overflowY: 'auto', minHeight: 0 }}>
 
-        {/* ── 外观设置 ── */}
         {activeTab === 'appearance' && (
-          <div className={sectionClass}>
-            <div className={sectionTitleClass}>🎨 外观 Appearance</div>
-
-            <div className="settings-item">
-              <label className="settings-label">主题 Theme</label>
-              <select
-                className="settings-select"
-                value={theme}
-                onChange={(e) => {
-                  const v = e.target.value as Theme
-                  setTheme(v)
-                  handleSave({ theme: v, font_size: fontSize, nickname, language })
-                }}
-              >
-                <option value="dark">深色 Dark</option>
-                <option value="light">浅色 Light</option>
-              </select>
-            </div>
-
-            <div className="settings-item">
-              <label className="settings-label">字体大小 Font Size</label>
-              <select
-                className="settings-select"
-                value={fontSize}
-                onChange={(e) => {
-                  const v = e.target.value as FontSize
-                  setFontSize(v)
-                  handleSave({ theme, font_size: v, nickname, language })
-                }}
-              >
-                <option value="small">小 Small</option>
-                <option value="medium">中 Medium</option>
-                <option value="large">大 Large</option>
-              </select>
-            </div>
-
-            <div className="settings-item">
-              <label className="settings-label">昵称 Nickname</label>
-              <input
-                type="text"
-                className="settings-input"
-                placeholder="输入昵称..."
-                maxLength={32}
-                value={nickname}
-                onChange={(e) => setNickname(e.target.value)}
-                onBlur={() => {
-                  onNicknameChange?.(nickname)
-                  handleSave({ theme, font_size: fontSize, nickname, language })
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    ;(e.target as HTMLInputElement).blur()
-                  }
-                }}
-              />
-            </div>
-
-            <div className="settings-item">
-              <label className="settings-label">语言 Language</label>
-              <select
-                className="settings-select"
-                value={language}
-                onChange={(e) => {
-                  const v = e.target.value as Language
-                  setLanguage(v)
-                  handleSave({ theme, font_size: fontSize, nickname, language: v })
-                }}
-              >
-                <option value="zh-CN">简体中文</option>
-                <option value="en">English</option>
-              </select>
-            </div>
-          </div>
+          <AppearanceTab showToast={showToast} onNicknameChange={onNicknameChange} onSavingChange={setSaving} />
         )}
 
-        {/* ── 会话管理 (ChatRooms) ── */}
         {activeTab === 'sessions' && (
-          <div className={sectionClass}>
-            <div className={sectionTitleClass}>💬 ChatRooms</div>
-            <p className="text-xs text-slate-500 mb-3">
-              所有对话都是 ChatRoom — 人↔Agent、Agent↔Agent 统一管理。
-            </p>
-
-            <div className="flex gap-2 mb-3">
-              <button
-                className="settings-action-btn"
-                onClick={fetchSessions}
-                disabled={sessionsLoading}
-              >
-                {sessionsLoading ? '加载中...' : '🔄 刷新'}
-              </button>
-              {selectedSession && (
-                <button
-                  className="settings-action-btn"
-                  onClick={() => { setSelectedSession(null); setSessionMessages([]) }}
-                >
-                  ← 返回列表
-                </button>
-              )}
-            </div>
-
-            {selectedSession ? (
-              /* ── 查看 ChatRoom 消息 ── */
-              <div>
-                <div className="flex items-center gap-2 mb-3 p-2 rounded-lg bg-slate-800/50">
-                  <span className={`text-xs px-1.5 py-0.5 rounded ${
-                    selectedSession.type === 'main'
-                      ? 'bg-emerald-900/50 text-emerald-400'
-                      : 'bg-indigo-900/50 text-indigo-400'
-                  }`}>
-                    {selectedSession.type === 'main' ? '👤 主会话' : '🤖 Agent'}
-                  </span>
-                  <span className="text-sm font-medium text-slate-200">
-                    {selectedSession.label}
-                  </span>
-                  {selectedSession.members && (
-                    <span className="text-xs text-slate-500">
-                      {selectedSession.members}
-                    </span>
-                  )}
-                  {selectedSession.type !== 'main' && (
-                    <span className={`ml-auto text-xs px-1.5 py-0.5 rounded ${
-                      selectedSession.running
-                        ? 'bg-green-900/50 text-green-400'
-                        : 'bg-slate-700 text-slate-400'
-                    }`}>
-                      {selectedSession.running ? '运行中' : '已完成'}
-                    </span>
-                  )}
-                </div>
-
-                {sessionMessagesLoading ? (
-                  <div className="text-center py-4 text-slate-500 text-sm">加载中...</div>
-                ) : sessionMessages.length === 0 ? (
-                  <div className="text-center py-4 text-slate-500 text-sm">暂无消息记录</div>
-                ) : (
-                  <div className="space-y-2 max-h-[400px] overflow-y-auto">
-                    {sessionMessages.map((msg, i) => (
-                      <div
-                        key={i}
-                        className={`p-2 rounded-lg text-sm ${
-                          msg.role === 'user'
-                            ? 'bg-blue-900/20 border-l-2 border-blue-500 ml-4'
-                            : msg.role === 'system'
-                            ? 'bg-yellow-900/20 border-l-2 border-yellow-500 mr-4 text-xs'
-                            : 'bg-slate-800/50 border-l-2 border-indigo-500 mr-4'
-                        }`}
-                      >
-                        <div className="text-xs text-slate-500 mb-1">
-                          {msg.role === 'user' ? '👤 User' : msg.role === 'system' ? '⚙️ System' : '🤖 Assistant'}
-                        </div>
-                        <div className="text-slate-300 whitespace-pre-wrap break-words" style={{maxHeight: '200px', overflowY: 'auto'}}>
-                          {msg.content}
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                )}
-              </div>
-            ) : (
-              /* ── ChatRoom 列表 ── */
-              sessionsLoading ? (
-                <div className="text-center py-6 text-slate-500 text-sm">加载中...</div>
-              ) : sessions.length === 0 ? (
-                <div className="text-center py-6 text-slate-500">
-                  <p className="text-2xl mb-2">📭</p>
-                  <p className="text-sm">暂无 ChatRoom</p>
-                </div>
-              ) : (
-                <div className="space-y-2">
-                  {sessions.map((s) => (
-                    <div
-                      key={s.id}
-                      className="p-3 rounded-lg bg-slate-800/50 hover:bg-slate-700/50 cursor-pointer transition-colors"
-                      onClick={() => setSelectedSession(s)}
-                    >
-                      <div className="flex items-center gap-2">
-                        <span className={`text-xs px-1.5 py-0.5 rounded ${
-                          s.type === 'main'
-                            ? 'bg-emerald-900/50 text-emerald-400'
-                            : 'bg-indigo-900/50 text-indigo-400'
-                        }`}>
-                          {s.type === 'main' ? '👤' : '🤖'}
-                        </span>
-                        <span className="text-sm font-medium text-slate-200">{s.label}</span>
-                        {s.members && (
-                          <span className="text-xs text-slate-500">{s.members}</span>
-                        )}
-                        {s.type !== 'main' && (
-                          <span className={`ml-auto text-xs px-1.5 py-0.5 rounded ${
-                            s.running
-                              ? 'bg-green-900/50 text-green-400'
-                              : 'bg-slate-700 text-slate-400'
-                          }`}>
-                            {s.running ? '运行中' : '已完成'}
-                          </span>
-                        )}
-                      </div>
-                      {s.preview && (
-                        <div className="text-xs text-slate-500 mt-1 truncate">{s.preview}</div>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              )
-            )}
-          </div>
+          <SessionsTab />
         )}
 
-        {/* ── 快捷指令 ── */}
         {activeTab === 'presets' && (
-          <div className={sectionClass}>
-            <div className={sectionTitleClass}>⚡ 快捷指令 Preset Commands</div>
-            <p className="text-xs text-slate-500 mb-3">
-              配置常用指令，在聊天输入框上方快速触发。最多 20 条。
-            </p>
-
-            {editingPreset ? (
-              /* ── 编辑/新增表单 ── */
-              <div className="preset-edit-form">
-                <div className="settings-item">
-                  <label className="settings-label">图标 Icon</label>
-                  <input
-                    type="text"
-                    className="settings-input"
-                    style={{ width: '60px' }}
-                    maxLength={4}
-                    value={editingPreset.icon}
-                    onChange={(e) => setEditingPreset({ ...editingPreset, icon: e.target.value })}
-                  />
-                </div>
-                <div className="settings-item">
-                  <label className="settings-label">名称 Label *</label>
-                  <input
-                    type="text"
-                    className="settings-input"
-                    placeholder="例如：代码审查"
-                    maxLength={20}
-                    value={editingPreset.label}
-                    onChange={(e) => setEditingPreset({ ...editingPreset, label: e.target.value })}
-                  />
-                </div>
-                <div className="settings-item">
-                  <label className="settings-label">内容 Content *</label>
-                  <textarea
-                    className="settings-input"
-                    style={{ minHeight: '80px', resize: 'vertical' }}
-                    placeholder="点击后发送的内容..."
-                    maxLength={2000}
-                    value={editingPreset.content}
-                    onChange={(e) => setEditingPreset({ ...editingPreset, content: e.target.value })}
-                  />
-                  <p className="text-xs text-slate-600 mt-1">{editingPreset.content.length}/2000</p>
-                </div>
-                <div className="settings-item">
-                  <label className="settings-label flex items-center gap-2">
-                    <input
-                      type="checkbox"
-                      checked={editingPreset.fill ?? false}
-                      onChange={(e) => setEditingPreset({ ...editingPreset, fill: e.target.checked })}
-                    />
-                    填充模式（填入输入框而非直接发送）
-                  </label>
-                </div>
-                <div className="flex gap-2 mt-3">
-                  <button
-                    className="settings-action-btn"
-                    onClick={() => handlePresetSave(editingPreset)}
-                    disabled={!editingPreset.label.trim() || !editingPreset.content.trim() || presetSaving}
-                  >
-                    {presetSaving ? '保存中...' : '💾 保存'}
-                  </button>
-                  <button
-                    className="settings-action-btn settings-action-danger"
-                    onClick={() => setEditingPreset(null)}
-                  >
-                    取消
-                  </button>
-                </div>
-              </div>
-            ) : (
-              /* ── 列表视图 ── */
-              <>
-                {presetList.length === 0 ? (
-                  <div className="text-center py-6 text-slate-500">
-                    <p className="text-2xl mb-2">📭</p>
-                    <p className="text-sm">暂无快捷指令</p>
-                  </div>
-                ) : (
-                  <div className="preset-list">
-                    {[...presetList].sort((a, b) => a.sort - b.sort).map((p) => (
-                      <div key={p.id} className="preset-item">
-                        <div className="preset-item-main">
-                          <span className="preset-item-icon">{p.icon || '⚡'}</span>
-                          <div className="preset-item-info">
-                            <span className="preset-item-label">{p.label}</span>
-                            <span className="preset-item-content">{p.content.length > 40 ? p.content.slice(0, 40) + '...' : p.content}</span>
-                          </div>
-                        </div>
-                        <div className="preset-item-actions">
-                          <button
-                            className="preset-action-btn"
-                            onClick={() => handlePresetMove(p.id, 'up')}
-                            title="上移"
-                            disabled={presetSaving}
-                          >↑</button>
-                          <button
-                            className="preset-action-btn"
-                            onClick={() => handlePresetMove(p.id, 'down')}
-                            title="下移"
-                            disabled={presetSaving}
-                          >↓</button>
-                          <button
-                            className="preset-action-btn"
-                            onClick={() => setEditingPreset({ ...p })}
-                            title="编辑"
-                            disabled={presetSaving}
-                          >✏️</button>
-                          <button
-                            className="preset-action-btn preset-action-delete"
-                            onClick={() => handlePresetDelete(p.id)}
-                            title="删除"
-                            disabled={presetSaving}
-                          >🗑️</button>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                )}
-                <button
-                  className="settings-action-btn w-full mt-3"
-                  onClick={handlePresetAdd}
-                  disabled={presetList.length >= 20 || presetSaving}
-                >
-                  ➕ 新增指令 {presetList.length > 0 ? `(${presetList.length}/20)` : ''}
-                </button>
-              </>
-            )}
-          </div>
+          <PresetsTab showToast={showToast} onPresetsChange={onPresetsChange} />
         )}
 
-        {/* ── LLM 设置 ── */}
         {activeTab === 'llm' && (
-          <div className={sectionClass}>
-            <div className={sectionTitleClass}>🧠 个人 LLM 配置</div>
-
-            {llmConfigLoading ? (
-              <div className="text-center py-6 text-slate-500 text-sm">加载中...</div>
-            ) : llmConfig && !llmConfig.is_global ? (
-              /* ── 个人配置：显示当前配置 + 模型切换 + 删除 ── */
-              <>
-                <div className="text-xs text-slate-400 mb-3">
-                  当前使用个人模型。可切换模型或删除配置以恢复系统默认。
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">提供商 Provider</label>
-                  <div className="text-sm text-slate-300">{providerLabel}</div>
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">Base URL</label>
-                  <div className="text-sm text-slate-400 font-mono break-all">{llmConfig.base_url}</div>
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">当前模型 Model</label>
-                  {llmConfig.models.length > 0 ? (
-                    <select
-                      className="settings-select"
-                      value={llmConfig.model}
-                      onChange={(e) => handleLLMSetModel(e.target.value)}
-                      disabled={llmSaving}
-                    >
-                      {llmConfig.models.map(m => (
-                        <option key={m} value={m}>{m}</option>
-                      ))}
-                    </select>
-                  ) : (
-                    <div className="text-sm text-slate-300">{llmConfig.model || '默认'}</div>
-                  )}
-                </div>
-
-                {llmError && <p className="text-xs text-red-400 mt-1 mb-2">{llmError}</p>}
-
-                <div className="flex gap-2 mt-3">
-                  <button
-                    className="settings-action-btn settings-action-danger"
-                    onClick={handleLLMDelete}
-                    disabled={llmSaving}
-                  >
-                    🗑️ 删除配置
-                  </button>
-                </div>
-
-              </>
-            ) : llmConfig && llmConfig.is_global ? (
-              /* ── 全局配置：显示当前模型 + 添加配置入口 ── */
-              <>
-                <div className="text-xs text-slate-400 mb-3">
-                  当前使用系统全局模型。配置个人 LLM 后可自由选择模型。
-                </div>
-
-                {llmConfig.model && (
-                  <div className="settings-item">
-                    <label className="settings-label">当前模型 Model</label>
-                    <div className="text-sm text-slate-300">{llmConfig.model}</div>
-                  </div>
-                )}
-
-                {llmError && <p className="text-xs text-red-400 mt-1 mb-2">{llmError}</p>}
-
-                <button
-                  className="settings-action-btn w-full mt-2"
-                  onClick={() => setLlmConfig(null)}
-                >
-                  ➕ 添加配置
-                </button>
-              </>
-            ) : (
-              /* ── 无配置：新增表单 ── */
-              <>
-                <div className="text-xs text-slate-400 mb-3">
-                  当前使用系统默认模型。配置个人 LLM 后可自由选择模型。
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">提供商 Provider</label>
-                  <select
-                    className="settings-select"
-                    value={llmFormProvider}
-                    onChange={(e) => setLlmFormProvider(e.target.value)}
-                  >
-                    {PROVIDER_OPTIONS.map(p => (
-                      <option key={p.value} value={p.value}>{p.label}</option>
-                    ))}
-                  </select>
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">Base URL *</label>
-                  <input
-                    type="text"
-                    className="settings-input"
-                    placeholder="例如: https://api.openai.com/v1"
-                    value={llmFormBaseUrl}
-                    onChange={(e) => setLlmFormBaseUrl(e.target.value)}
-                  />
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">API Key *</label>
-                  <input
-                    type="password"
-                    className="settings-input"
-                    placeholder="sk-..."
-                    value={llmFormApiKey}
-                    onChange={(e) => setLlmFormApiKey(e.target.value)}
-                  />
-                  <p className="text-xs text-slate-600 mt-1">⚠️ API Key 仅存储在服务端，不会返回到前端</p>
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">模型 Model</label>
-                  <input
-                    type="text"
-                    className="settings-input"
-                    placeholder="例如: gpt-4o, claude-sonnet-4-20250514（可选，默认用提供商推荐模型）"
-                    value={llmFormModel}
-                    onChange={(e) => setLlmFormModel(e.target.value)}
-                  />
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">Max Output Tokens</label>
-                  <input
-                    type="number"
-                    className="settings-input"
-                    placeholder="0 = use default"
-                    value={llmFormMaxOutputTokens}
-                    onChange={e => setLlmFormMaxOutputTokens(e.target.value)}
-                    min={0}
-                    step={256}
-                  />
-                </div>
-
-                <div className="settings-item">
-                  <label className="settings-label">Thinking Mode</label>
-                  <select
-                    className="settings-select"
-                    value={llmFormThinkingMode}
-                    onChange={e => setLlmFormThinkingMode(e.target.value)}
-                  >
-                    <option value="auto">Auto</option>
-                    <option value="enabled">Enabled</option>
-                    <option value="disabled">Disabled</option>
-                  </select>
-                </div>
-
-                {llmError && <p className="text-xs text-red-400 mt-1 mb-2">{llmError}</p>}
-
-                <button
-                  className="settings-action-btn w-full mt-2"
-                  onClick={handleLLMAdd}
-                  disabled={llmSaving}
-                >
-                  {llmSaving ? '保存中...' : '💾 保存配置'}
-                </button>
-              </>
-            )}
-
-            {/* ── Max Context 设置（独立于 LLM 配置） ── */}
-            {!llmConfigLoading && (
-              <div className="settings-item mt-4 border-t border-slate-700/30 pt-4">
-                <label className="settings-label">最大上下文 Max Context</label>
-                <div className="flex items-center gap-2">
-                  <input
-                    type="number"
-                    className="settings-input flex-1"
-                    min={0}
-                    step={1000}
-                    value={llmMaxContext || 0}
-                    onChange={(e) => setLlmMaxContext(parseInt(e.target.value) || 0)}
-                    placeholder="0 = 使用系统默认"
-                  />
-                  <button
-                    className="settings-action-btn"
-                    onClick={handleLLMMaxContextSave}
-                    disabled={llmMaxContextSaving}
-                  >
-                    {llmMaxContextSaving ? '⏳' : '💾'} 保存
-                  </button>
-                </div>
-                <div className="text-[11px] text-slate-500 mt-1">
-                  Token 数量，0 表示使用系统默认值。值越大，可用的对话上下文越长，但消耗的 Token 越多。
-                </div>
-              </div>
-            )}
-
-          </div>
+          <LLMTab showToast={showToast} />
         )}
 
-        {/* ── Runner 设置 ── */}
-        {activeTab === 'runner' && <RunnerPanel />}
+        {activeTab === 'runner' && <RunnerTab />}
 
-        {/* ── Agent 市场 ── */}
         {activeTab === 'market' && (
-          <div className={sectionClass}>
-            <div className={sectionTitleClass}>🏪 Agent 市场</div>
-            <div className="market-tab-bar">
-              <button
-                className={`market-tab ${marketType === 'agent' ? 'active' : ''}`}
-                onClick={() => { setMarketType('agent'); setMarketSubTab('browse'); }}
-              >
-                🤖 Agent
-              </button>
-              <button
-                className={`market-tab ${marketType === 'skill' ? 'active' : ''}`}
-                onClick={() => { setMarketType('skill'); setMarketSubTab('browse'); }}
-              >
-                🛠️ Skill
-              </button>
-            </div>
-            {/* Sub tabs: browse / mine */}
-            <div className="market-sub-tab-bar">
-              <button
-                className={`market-tab market-sub-tab ${marketSubTab === 'browse' ? 'active' : ''}`}
-                onClick={() => setMarketSubTab('browse')}
-              >
-                📦 市场
-              </button>
-              <button
-                className={`market-tab market-sub-tab ${marketSubTab === 'mine' ? 'active' : ''}`}
-                onClick={() => setMarketSubTab('mine')}
-              >
-                📋 我的
-              </button>
-            </div>
-
-            {marketSubTab === 'browse' && (
-              marketLoading ? (
-                <div className="text-center py-8 text-slate-500">
-                  <div className="market-spinner" />
-                  <p className="text-xs mt-2">加载中...</p>
-                </div>
-              ) : marketEntries.length === 0 ? (
-                <div className="text-center py-8 text-slate-500">
-                  <p className="text-3xl mb-3">📭</p>
-                  <p className="text-sm">暂无可用条目</p>
-                </div>
-              ) : (
-                <div className="market-entry-list">
-                  {marketEntries.map(entry => (
-                    <div key={entry.id} className="market-entry">
-                      <div className="market-entry-header">
-                        <div className="market-entry-info">
-                          <span className="market-entry-name">{entry.name}</span>
-                          <span className="market-entry-author">by {entry.author}</span>
-                        </div>
-                        {entry.installed ? (
-                          <button className="market-uninstall-btn" onClick={() => handleUninstall(entry)}>
-                            卸载
-                          </button>
-                        ) : (
-                          <button className="market-install-btn" onClick={() => handleInstall(entry)}>
-                            安装
-                          </button>
-                        )}
-                      </div>
-                      {entry.description && (
-                        <p className="market-entry-desc">{entry.description}</p>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              )
-            )}
-
-            {marketSubTab === 'mine' && (
-              marketLoading ? (
-                <div className="text-center py-8 text-slate-500">
-                  <div className="market-spinner" />
-                  <p className="text-xs mt-2">加载中...</p>
-                </div>
-              ) : myMarketEntries.length === 0 ? (
-                <div className="text-center py-8 text-slate-500">
-                  <p className="text-3xl mb-3">📭</p>
-                  <p className="text-sm">暂无自己的{marketType === 'skill' ? ' Skill' : ' Agent'}</p>
-                </div>
-              ) : (
-                <div className="market-entry-list">
-                  {myMarketEntries.map(entry => (
-                    <div key={entry.name} className="market-entry">
-                      <div className="market-entry-header">
-                        <div className="market-entry-info">
-                          <span className="market-entry-name">{entry.name}</span>
-                          <span className={`market-entry-status ${entry.published ? 'published' : 'unpublished'}`}>
-                            {entry.published ? '✅ 已上架' : '⚪ 未上架'}
-                          </span>
-                        </div>
-                        {entry.published ? (
-                          <button className="market-unpublish-btn" onClick={() => handleUnpublish(entry)}>
-                            下架
-                          </button>
-                        ) : (
-                          <button className="market-install-btn" onClick={() => handlePublish(entry)}>
-                            上架
-                          </button>
-                        )}
-                      </div>
-                      {entry.description && (
-                        <p className="market-entry-desc">{entry.description}</p>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              )
-            )}
-          </div>
+          <MarketTab />
         )}
+
         </div>{/* end scrollable content */}
       </div>
       {/* Toast notification */}
