+++ web/src/components/settings/RunnerTab.tsx
@@ -1,0 +1,5 @@
+import RunnerPanel from '../RunnerPanel'
+
+export default function RunnerTab() {
+  return <RunnerPanel />
+}
