+++ web/src/styles/codeblock.css
@@ -1,0 +1,215 @@
+/* ========================================================================== */
+/* Code Block Styles                                                         */
+/* ========================================================================== */
+
+/* ============ Code blocks — see "Enhanced Code Block" section below for main styles ============ */
+
+/* Inline code (dark mode default) */
+.xbot-inline-code {
+  background: #1e293b;
+  color: #e2e8f0;
+  padding: 2px 6px;
+  border-radius: 4px;
+  font-size: 0.9em;
+  word-break: break-word;
+  overflow-wrap: anywhere;
+}
+
+/* User message bubble inline code: override for blue background */
+.bg-blue-600 .xbot-inline-code,
+.bg-blue-700 .xbot-inline-code {
+  background: rgba(255,255,255,0.25);
+  color: #fff;
+}
+
+/* ---- Enhanced Code Block (dark) ---- */
+
+/* ========================================================================== */
+/* Enhanced Code Block                                                         */
+/* ========================================================================== */
+
+.xbot-codeblock {
+  position: relative;
+  border-radius: 10px;
+  overflow: hidden;
+  border: 1px solid rgba(51, 65, 85, 0.5);
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
+  transition: border-color 0.2s ease;
+  margin: 0.5em 0;
+}
+
+.xbot-codeblock:hover {
+  border-color: rgba(71, 85, 105, 0.7);
+}
+
+.xbot-codeblock-header {
+  background: linear-gradient(135deg, #1e293b, #0f172a);
+  padding: 6px 12px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border-bottom: 1px solid rgba(51, 65, 85, 0.3);
+  font-size: 12px;
+  color: #8b949e;
+}
+
+.xbot-codeblock-copy {
+  background: rgba(71, 85, 105, 0.3);
+  border: none;
+  color: #94a3b8;
+  padding: 2px 8px;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 11px;
+  transition: all 0.15s ease;
+}
+
+.xbot-codeblock-copy:hover {
+  background: rgba(71, 85, 105, 0.5);
+  color: #e2e8f0;
+}
+
+.xbot-codeblock-pre {
+  margin: 0;
+  padding: 12px 16px;
+  overflow-x: auto;
+  background: #0d1117;
+  font-size: 13px;
+  line-height: 1.6;
+}
+
+.xbot-codeblock-pre code {
+  background: transparent;
+  padding: 0;
+  font-size: 0.875em;
+  border-radius: 0;
+}
+
+/* Inline code enhancement */
+.xbot-inline-code {
+  background: linear-gradient(135deg, rgba(30, 41, 59, 0.8), rgba(15, 23, 42, 0.6));
+  padding: 1px 6px;
+  border-radius: 4px;
+  font-size: 0.88em;
+  border: 1px solid rgba(51, 65, 85, 0.3);
+  color: #93c5fd;
+  word-break: break-word;
+  overflow-wrap: anywhere;
+}
+
+/* ---- Light Theme Overrides ---- */
+
+/* Light mode hljs override (github-light key colors) */
+[data-theme="light"] .hljs {
+  color: #24292e;
+  background: transparent;
+}
+[data-theme="light"] .hljs-comment,
+[data-theme="light"] .hljs-quote {
+  color: #6a737d;
+  font-style: italic;
+}
+[data-theme="light"] .hljs-keyword,
+[data-theme="light"] .hljs-selector-tag,
+[data-theme="light"] .hljs-type {
+  color: #d73a49;
+}
+[data-theme="light"] .hljs-string,
+[data-theme="light"] .hljs-addition {
+  color: #22863a;
+}
+[data-theme="light"] .hljs-number,
+[data-theme="light"] .hljs-literal {
+  color: #005cc5;
+}
+[data-theme="light"] .hljs-built_in,
+[data-theme="light"] .hljs-builtin-name {
+  color: #e36209;
+}
+[data-theme="light"] .hljs-function .hljs-title,
+[data-theme="light"] .hljs-title.function_ {
+  color: #6f42c1;
+}
+[data-theme="light"] .hljs-variable,
+[data-theme="light"] .hljs-attr {
+  color: #e36209;
+}
+[data-theme="light"] .hljs-params {
+  color: #24292e;
+}
+[data-theme="light"] .hljs-meta {
+  color: #735c0f;
+}
+[data-theme="light"] .hljs-deletion {
+  color: #b31d28;
+  background: #ffeef0;
+}
+[data-theme="light"] .hljs-addition {
+  color: #22863a;
+  background: #f0fff4;
+}
+[data-theme="light"] .hljs-symbol,
+[data-theme="light"] .hljs-bullet {
+  color: #005cc5;
+}
+[data-theme="light"] .hljs-regexp {
+  color: #032f62;
+}
+[data-theme="light"] .hljs-selector-class {
+  color: #6f42c1;
+}
+[data-theme="light"] .hljs-selector-id {
+  color: #005cc5;
+}
+[data-theme="light"] .hljs-name,
+[data-theme="light"] .hljs-tag {
+  color: #22863a;
+}
+[data-theme="light"] .hljs-attribute {
+  color: #005cc5;
+}
+[data-theme="light"] .hljs-section {
+  color: #005cc5;
+  font-weight: bold;
+}
+[data-theme="light"] .hljs-link {
+  color: #032f62;
+  text-decoration: underline;
+}
+
+
+/* Light mode overrides for code blocks */
+[data-theme="light"] .xbot-codeblock {
+  border-color: #d1cdc7;
+  box-shadow: 0 2px 8px rgba(28, 25, 23, 0.08);
+}
+
+[data-theme="light"] .xbot-codeblock:hover {
+  border-color: #bbb7b0;
+}
+
+[data-theme="light"] .xbot-codeblock-header {
+  background: linear-gradient(135deg, #e2dfda, #ddd9d4);
+  color: #57534e;
+  border-bottom-color: #d1cdc7;
+}
+
+[data-theme="light"] .xbot-codeblock-pre {
+  background: #edeae5;
+}
+
+[data-theme="light"] .xbot-codeblock-copy {
+  color: #78716c;
+  background: rgba(28, 25, 23, 0.06);
+}
+
+[data-theme="light"] .xbot-codeblock-copy:hover {
+  color: #3d3833;
+  background: rgba(28, 25, 23, 0.12);
+}
+
+[data-theme="light"] .xbot-inline-code {
+  background: #e2dfda;
+  color: #1e293b;
+  border-color: #d1cdc7;
+}
