+++ web/src/styles/login.css
@@ -1,0 +1,7 @@
+/* ========================================================================== */
+/* Login Page Styles                                                         */
+/* ========================================================================== */
+
+/* Login-specific styles are global light theme overrides for input fields. */
+/* See base.css for [data-theme=light] input.bg-slate-700, */
+/* select.bg-slate-700, and placeholder-slate-400::placeholder rules. */
