+++ web/src/styles/progress.css
@@ -1,0 +1,260 @@
+/* ========================================================================== */
+/* SubAgent Tree Styles                                                      */
+/* ========================================================================== */
+
+/* ========================================================================== */
+/* SubAgent Tree Styles                                                       */
+/* ========================================================================== */
+
+@keyframes subagent-pulse {
+  0%, 100% { opacity: 1; }
+  50% { opacity: 0.4; }
+}
+
+@keyframes subagent-spin {
+  from { transform: rotate(0deg); }
+  to { transform: rotate(360deg); }
+}
+
+.subagent-pulse {
+  animation: subagent-pulse 1.5s ease-in-out infinite;
+  display: inline-block;
+}
+
+.subagent-spin {
+  animation: subagent-spin 2s linear infinite;
+  display: inline-block;
+}
+
+.subagent-tree {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+}
+
+.subagent-node {
+  position: relative;
+}
+
+.subagent-node-running {
+  background: rgba(59, 130, 246, 0.08);
+  border-left: 2px solid rgba(59, 130, 246, 0.4);
+}
+
+.subagent-node-pending {
+  background: rgba(148, 163, 184, 0.05);
+  border-left: 2px solid rgba(148, 163, 184, 0.3);
+}
+
+/* ========================================================================== */
+/* SubAgent Animations                                                       */
+/* ========================================================================== */
+
+/* ========================================================================== */
+/* SubAgent Animations                                                         */
+/* ========================================================================== */
+
+.subagent-spin {
+  display: inline-block;
+  animation: subagent-rotate 2s linear infinite;
+}
+
+@keyframes subagent-rotate {
+  from { transform: rotate(0deg); }
+  to { transform: rotate(360deg); }
+}
+
+.subagent-pulse {
+  display: inline-block;
+  animation: subagent-pulse-anim 1.5s ease-in-out infinite;
+}
+
+@keyframes subagent-pulse-anim {
+  0%, 100% { opacity: 0.5; }
+  50% { opacity: 1; }
+}
+
+.subagent-node-running {
+  background: rgba(59, 130, 246, 0.08);
+  border-left: 2px solid rgba(59, 130, 246, 0.4);
+}
+
+.subagent-node-pending {
+  border-left: 2px solid rgba(100, 116, 139, 0.3);
+}
+
+
+/* ========================================================================== */
+/* Enhanced Thinking Animation (Orb)                                         */
+/* ========================================================================== */
+
+/* ========================================================================== */
+/* Enhanced Thinking Animation (Orb)                                           */
+/* ========================================================================== */
+
+.thinking-orb {
+  position: relative;
+  width: 28px;
+  height: 28px;
+  flex-shrink: 0;
+}
+
+.thinking-orb-sm {
+  width: 22px;
+  height: 22px;
+}
+
+.thinking-orb-core {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 8px;
+  height: 8px;
+  background: radial-gradient(circle, #60a5fa, #3b82f6);
+  border-radius: 50%;
+  transform: translate(-50%, -50%);
+  box-shadow: 0 0 8px rgba(59, 130, 246, 0.6);
+}
+
+.thinking-orb-sm .thinking-orb-core {
+  width: 6px;
+  height: 6px;
+}
+
+.thinking-orb-ring {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  border-radius: 50%;
+  border: 1.5px solid transparent;
+  transform: translate(-50%, -50%);
+}
+
+.thinking-orb-ring-1 {
+  width: 100%;
+  height: 100%;
+  border-top-color: rgba(96, 165, 250, 0.8);
+  border-right-color: rgba(96, 165, 250, 0.3);
+  animation: orb-spin 1.2s linear infinite;
+}
+
+.thinking-orb-ring-2 {
+  width: 80%;
+  height: 80%;
+  border-bottom-color: rgba(34, 197, 94, 0.6);
+  border-left-color: rgba(34, 197, 94, 0.2);
+  animation: orb-spin 1.8s linear infinite reverse;
+}
+
+.thinking-orb-ring-3 {
+  width: 60%;
+  height: 60%;
+  border-top-color: rgba(168, 85, 247, 0.5);
+  animation: orb-spin 2.4s linear infinite;
+}
+
+@keyframes orb-spin {
+  from { transform: translate(-50%, -50%) rotate(0deg); }
+  to { transform: translate(-50%, -50%) rotate(360deg); }
+}
+
+/* ========================================================================== */
+/* Enhanced Thinking Dots                                                    */
+/* ========================================================================== */
+
+/* ========================================================================== */
+/* Enhanced Thinking Dots                                                      */
+/* ========================================================================== */
+
+.thinking-dots {
+  display: flex;
+  gap: 4px;
+  align-items: center;
+}
+
+.thinking-dot {
+  width: 5px;
+  height: 5px;
+  border-radius: 50%;
+  background: #64748b;
+  animation: dot-wave 1.4s ease-in-out infinite;
+}
+
+@keyframes dot-wave {
+  0%, 80%, 100% {
+    transform: scale(0.6);
+    opacity: 0.4;
+  }
+  40% {
+    transform: scale(1);
+    opacity: 1;
+    background: #60a5fa;
+    box-shadow: 0 0 6px rgba(96, 165, 250, 0.5);
+  }
+}
+
+/* ========================================================================== */
+/* Enhanced Tool Pulse Animation                                             */
+/* ========================================================================== */
+
+/* ========================================================================== */
+/* Enhanced Tool Pulse Animation                                               */
+/* ========================================================================== */
+
+.tool-pulse {
+  display: inline-block;
+  animation: tool-glow 2s ease-in-out infinite;
+}
+
+@keyframes tool-glow {
+  0%, 100% { opacity: 0.7; transform: scale(1); }
+  50% { opacity: 1; transform: scale(1.15); box-shadow: 0 0 4px rgba(59, 130, 246, 0.3); }
+}
+
+/* ---- Light Theme Overrides ---- */
+
+/* Thinking orb — softer glow for light backgrounds */
+[data-theme="light"] .thinking-orb-core {
+  background: radial-gradient(circle, #3b82f6, #2563eb);
+  box-shadow: 0 0 6px rgba(59, 130, 246, 0.4);
+}
+[data-theme="light"] .thinking-orb-ring-1 {
+  border-top-color: rgba(59, 130, 246, 0.7);
+  border-right-color: rgba(59, 130, 246, 0.2);
+}
+[data-theme="light"] .thinking-orb-ring-2 {
+  border-bottom-color: rgba(22, 163, 74, 0.5);
+  border-left-color: rgba(22, 163, 74, 0.15);
+}
+[data-theme="light"] .thinking-orb-ring-3 {
+  border-top-color: rgba(126, 34, 206, 0.4);
+}
+
+[data-theme="light"] .thinking-dot {
+  background: #a8a29e;
+}
+[data-theme="light"] .thinking-dot {
+  animation-name: dot-wave-light;
+}
+@keyframes dot-wave-light {
+  0%, 80%, 100% {
+    transform: scale(0.6);
+    opacity: 0.4;
+    background: #a8a29e;
+  }
+  40% {
+    transform: scale(1);
+    opacity: 1;
+    background: #2563eb;
+    box-shadow: 0 0 4px rgba(37, 99, 235, 0.3);
+  }
+}
+
+/* SubAgent nodes */
+[data-theme="light"] .subagent-node-running {
+  background: rgba(37, 99, 235, 0.06);
+  border-left-color: rgba(37, 99, 235, 0.3);
+}
+[data-theme="light"] .subagent-node-pending {
+  border-left-color: rgba(168, 162, 158, 0.4);
+}
