+++ web/src/styles/sidebar.css
@@ -1,0 +1,87 @@
+/* ========================================================================== */
+/* ChatSidebar Component                                                    */
+/* ========================================================================== */
+
+/* ========================================================================== */
+/* ChatSidebar Component                                                      */
+/* ========================================================================== */
+
+.sidebar-btn {
+  font-size: 12px;
+  padding: 2px 6px;
+  border-radius: 4px;
+  background: transparent;
+  border: none;
+  color: #94a3b8;
+  cursor: pointer;
+  transition: all 0.15s;
+}
+.sidebar-btn:hover {
+  background: #334155;
+  color: #e2e8f0;
+}
+
+.sidebar-item {
+  margin: 2px 4px;
+  padding: 6px 8px;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: background 0.15s;
+  border-left: 2px solid transparent;
+}
+.sidebar-item:hover {
+  background: rgba(30, 41, 59, 0.5);
+}
+.sidebar-item-active {
+  background: rgba(67, 56, 202, 0.15);
+  border-left-color: #6366f1;
+}
+
+.sidebar-delete-btn {
+  display: none;
+  font-size: 10px;
+  color: #475569;
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 0 2px;
+  transition: color 0.15s;
+}
+.sidebar-item:hover .sidebar-delete-btn {
+  display: inline-block;
+}
+.sidebar-delete-btn:hover {
+  color: #f87171;
+}
+
+.sidebar-rename-input {
+  flex: 1;
+  background: #1e293b;
+  border: 1px solid #475569;
+  border-radius: 4px;
+  color: #e2e8f0;
+  font-size: 12px;
+  padding: 1px 4px;
+  outline: none;
+}
+.sidebar-rename-input:focus {
+  border-color: #6366f1;
+}
+
+.sidebar-refresh-btn {
+  font-size: 10px;
+  color: #64748b;
+  background: none;
+  border: none;
+  cursor: pointer;
+  width: 100%;
+  text-align: center;
+  padding: 4px;
+  transition: color 0.15s;
+}
+.sidebar-refresh-btn:hover {
+  color: #cbd5e1;
+}
+.sidebar-refresh-btn:disabled {
+  cursor: default;
+}
